+++ addons/pvr.mythtv.cmyth/src/cppmyth/MythFile.cpp
@@ -55,7 +55,7 @@ void MythFile::UpdateLength(unsigned long long length)
   (void)retval;
 }
 
-int MythFile::Read(void *buffer, unsigned long length)
+int MythFile::Read(void *buffer, unsigned int length)
 {
   int bytesRead;
   bytesRead = cmyth_file_read(*m_file_t, static_cast< char * >(buffer), length);
